refactor(steps): delete auto-detect DetectServicesStep
PR 2 wired pv.yml-driven service binding while keeping the legacy
auto-detect path as a fallback. PR 4 shipped `pv init` so users on
existing projects have a migration tool. With pv.yml now the
contract, the legacy heuristic-based binding is dead:

- internal/automation/steps/detect_services.go deleted (helpers
  moved to apply_pvyml_services.go in the prior commit).
- detect_services_test.go deleted.
- cmd/link.go pipeline entry removed.
- pipeline.go LookupGate "detect_services" case removed.
- settings.go Automation.DetectServices field removed (old
  settings.yml files with the field deserialize cleanly — Go
  ignores unknown YAML fields).
