+++ cmd/link.go
@@ -152,7 +152,6 @@ pv link --name=myapp ~/Code/myapp`,
 			&steps.GenerateCaddyfileStep{},
 			&steps.GenerateTLSCertStep{},
 			&steps.ApplyPvYmlServicesStep{},
-			&steps.DetectServicesStep{},
 			&laravel.DetectServicesStep{},
 			&steps.ApplyPvYmlEnvStep{},
 			&laravel.SetAppURLStep{},
